pyproject.toml
[tool.poetry]
name = "breach-bot"
version = "0.1.0"
description = ""
authors = ["chahan69 <chahan69@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.*"
environ-config = "^21.2.0"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
